Fix illegal memory access
We were accessing an array index that was outside the bounds we defined.
I used this article to help me: http://jblevins.org/log/double-free

It had me re-compile the source using an additional option: -fbounds-check
After I did that, when I ran the code, it gave me a helpful message
about where the illegal access was taking place and what the requested
index was.