Hooked up funcionality to read in parameters from file
Added the ability to read in the variables from the file.  The
input file must have everything in the order that is listed in
code.  You can include comments at the end of the line.  However
you can't have comments on their own line by themselves.

Tried to compile in the coefficient subroutine to make sure it
was working.  Started getting a bunch of compiler errors.  Decided
against it.

got the Makefile working for this assignment.  If you get the coefficent
sub routine working, you'll have to swap out the "MOD" statements in
the makefile and then make sure the name is correct.  I just chose
utility b/c I thought that was a good name

when trying to get the coefficient subroutine working, i copied all
of it into a new file and renamed it utility.f90.  currently is
not used by anything